terminado backend carta

+++ app/src/main/java/g5/app/CartaRepository.java
@@ -3,7 +3,7 @@ package g5.app;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import java.util.*;
 
-public interface CartaRepository extends MongoRepository<Carta,Integer>{
+public interface CartaRepository extends MongoRepository<Carta,Long>{
 
 	Optional<Carta> findBynombreRestaurante(String nombre);
